+++ example/example/foo/urls.py
@@ -1,8 +1,16 @@
 from django.conf.urls import patterns, url
 
+from foo.views import FooDetailView
+
 urlpatterns = patterns('foo.views',
     # Listing URL
     url(r'^$', view='browse', name='foo.browse'),
+
+    url(r'^logged-in/$', view='logged_in', name='foo.logged_in'),
+
+    # Class based detail view URL
+    url(r'^class-based/(?P<slug>[\w\-\_\.\,]+)/$', FooDetailView.as_view(),
+        name='foo.class-based.detail'),
 
     # Detail URL
     url(r'^(?P<slug>[\w\-\_\.\,]+)/$', view='detail', name='foo.detail'),