+++ src/ska/generate_signed_url.py
@@ -1,3 +1,5 @@
+from __future__ import print_function
+
 __title__ = 'ska.generate_signed_url'
 __author__ = 'Artur Barseghyan'
 __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
@@ -5,7 +7,6 @@ __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('main',)
 
 import argparse
-from six import print_
 
 from ska import sign_url
 
@@ -70,9 +71,9 @@ def main():
 
     try:
         signed_url = sign_url(**kwargs)
-        print_(signed_url)
+        print(signed_url)
     except Exception as e:
-        print_(e)
+        print(e)
 
 if __name__ == "__main__":
     main()